+++ src/pointofsale/controllers/components/CardIngredientWhitManagerController.java
@@ -56,7 +56,7 @@ public class CardIngredientWhitManagerController implements ActionListener,Focus
         }
     }
 
-    private boolean validateRequest(Integer quantity) {
+    private boolean validateRequest(Long quantity) {
         if (quantity == 0) {
             return false;
         } else {
@@ -73,7 +73,7 @@ public class CardIngredientWhitManagerController implements ActionListener,Focus
     @Override
     public void actionPerformed(ActionEvent ae) {
         Object source = ae.getSource();
-        Integer quantity = Integer.valueOf(this.view.txtQuantity.getText());
+        Long quantity = Long.valueOf(this.view.txtQuantity.getText());
         if (source == this.view.btnAdd) {
             if (validateRequest(quantity)) {
                 this.ingredient.setQuantity(quantity);
@@ -85,12 +85,12 @@ public class CardIngredientWhitManagerController implements ActionListener,Focus
         }
 
         if (source == this.view.btnPlus) {
-            Integer sum = quantity + 1;
+            Long sum = quantity + 1;
             this.view.txtQuantity.setText(String.valueOf(sum));
         }
 
         if (source == this.view.btnLess) {
-            Integer less = quantity - 1;
+            Long less = quantity - 1;
             this.view.txtQuantity.setText(String.valueOf(less));
         }
