+++ src/pointofsale/dao/BillDao.java
@@ -12,6 +12,6 @@ import pointofsale.objects.Bill;
  * @author dragonyte
  */
 public interface BillDao extends Dao<Bill, Long>{
-    List<Bill> selectByEvent(Integer event_id);
-    Bill selectCollectEvent(Integer event_id);
+    List<Bill> selectByEvent(Long event_id);
+    Bill selectCollectEvent(Long event_id);
 }
